backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "insightxpert"
version = "0.1.0"
description = "AI data analyst for Techfest IIT Bombay Leadership Analytics Challenge"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0",
    "chromadb>=0.5.0",
    "google-genai>=1.0.0",
    "ollama>=0.4.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "sse-starlette>=2.0",
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/insightxpert"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
